fix saving documents
