+++ app/src/context/ClaimsStore.tsx
@@ -93,14 +93,14 @@ export const ClaimsProvider: React.FC<{
   );
 
   const addClaim = async (
-    claimId: ClaimType,
+    claimType: ClaimType,
     value: AddClaim_Value,
     files: string[],
     verified?: boolean
   ) => {
     // This is a mock function.
     // In the future we will send this data off to an api to be verified
-    console.log("making a claim", claimId, value, files);
+    console.log("making a claim", claimType, value, files);
 
     await new Promise((resolve) =>
       setTimeout(() => {
@@ -109,10 +109,12 @@ export const ClaimsProvider: React.FC<{
     );
 
     setClaimData((prevClaimData) => {
-      const otherClaimData = prevClaimData.filter((cd) => cd.type !== claimId);
+      const otherClaimData = prevClaimData.filter(
+        (cd) => cd.type !== claimType
+      );
       const updatedClaims = [
         ...otherClaimData,
-        { type: claimId, value, verified }
+        { type: claimType, value, files, verified }
       ];
       claimsLocalStorage.save(updatedClaims);
       return updatedClaims;
