+++ app/src/types/claim.ts
@@ -29,12 +29,14 @@ export type Claim = {
 export type ClaimData = {
   type: ClaimType;
   value: any; // each claim type has it own value type. todo - make generic
+  files: string[];
   verified?: boolean;
 };
 
 export type ClaimWithValue = Claim & {
   value: any; // each claim type has it own value type. todo - make generic
   verified?: boolean;
+  files: string[];
 };
 
 export type ClaimRequestParams = {
